+++ CHANGELOG.md
@@ -45,6 +45,7 @@ Misc:
 - Specify max heap size for Java tools [#2366](https://github.com/sider/runners/pull/2366)
 - **RuboCop** Add 3rd-party plugins [#2370](https://github.com/sider/runners/pull/2370)
 - **RuboCop** Fix links [#2371](https://github.com/sider/runners/pull/2371)
+- **CoffeeLint** Drop support for older versions [#2372](https://github.com/sider/runners/pull/2372)
 
 ## 0.47.0